+++ CHANGES.rst
@@ -41,6 +41,7 @@ v2.0.0 (xxxx-xx-xx)
 - Modify ``flatten()`` to be shallow by default, remove callback option, and add ``is_deep`` option. (**breaking change**)
 - Modify ``invert()`` to accept ``multivalue`` argument.
 - Modify ``is_number()`` to return ``False`` for boolean ``True`` and ``False``. (**breaking change**)
+- Modify ``result()`` to accept ``default`` argument.
 - Move files in ``pydash/api/`` to ``pydash/``. (**possible breaking change**)
 - Move predicate functions from ``pydash.api.objects`` to ``pydash.api.predicates``. (**possible breaking change**)
 - Rename ``create_callback()`` to ``iteratee``. (**breaking change**)